Add edit and delete functionality for students, update redirects

+++ pages/siswa/hapus_siswa.php
@@ -1,18 +1,20 @@
 <?php 
-include "../../inc/koneksi.php";
-$id=$_GET['id'];
-$q_hapus_siswa="DELETE FROM tbl_siswa WHERE nis='$id'";
-$hapus_siswa=mysqli_query($connect,$q_hapus_siswa);
-if($hapus_siswa){
-  echo "Data telah terhapus";
-  ?>
-  <script>
-    alert('Data telah terhapus')
-  </script>
-  <?php
-  header("Location: /21_michael_adminlte/?page=data_siswa");
-} else{
-  echo "Data gagal terhapus";
+// include "../../inc/koneksi.php";
+if (!isset($_GET['id']) || empty($_GET['id'])) {
+  die("ID tidak ditemukan.");
 }
-echo $id;
+
+$id = mysqli_real_escape_string($connect, $_GET['id']);
+$q_hapus_siswa = "DELETE FROM tbl_siswa WHERE nis='$id'";
+$hapus_siswa = mysqli_query($connect, $q_hapus_siswa);
+
+if ($hapus_siswa) {
+  echo "<script>
+      alert('Data telah terhapus');
+      window.location.href = '/AdminLTE/?page=data_siswa';
+  </script>";
+  exit();
+} else {
+  echo "Data gagal terhapus: " . mysqli_error($connect);
+}
 ?>
